start/stop command

+++ app/Console/Command/StopCommand.php
@@ -2,14 +2,9 @@
 namespace App\Console\Command;
 
 use App\Console\Command;
-use App\Service\Server;
-use Symfony\Component\Console\Command\LockableTrait;
-use Symfony\Component\Console\Input\InputArgument;
 
 class StopCommand extends Command
 {
-    use LockableTrait;
-
     protected $name = 'server:stop';
 
     protected $description = 'description';
@@ -23,9 +18,13 @@ class StopCommand extends Command
 
     public function handle()
     {
-        $pid = file_get_contents(getcwd() . '/daemon.pid');
-
-        posix_kill($pid, SIGINT);
+        $pidFile = getcwd() . '/daemon.pid';
+        if (file_exists($pidFile))
+        {
+            $pid = file_get_contents(getcwd() . '/daemon.pid');
+            posix_kill($pid, SIGINT);
+            unlink($pidFile);
+        }
     }
 
 }
